python_code/requirements.txt
pandas
python-dotenv
pinecone
firebase-admin
google-cloud-storage
mlxtend
openai
openai==1.61.1
runpod==1.7.7